Prevent segfault when freeing objects with DEBUG_LOG_GC enabled.
Note that some objects reference other objects which might be created before them.  This results in them being in front of them in the GC list, hence when printing details during freeObjects(), the front object will already be destroyed and invalid.  The particular case discovered was String names for Functions.